pyproject.toml
[tool.poetry]
name = "tm-map-sessions-v1"
version = "0.1.0"
description = ""
authors = ["smbrine <smbrine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.64.0"
grpcio-tools = "^1.64.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pydantic = "^2.7.2"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
pylint = "^3.2.2"
ruff = "^0.4.7"
grpcio-reflection = "^1.64.0"


[tool.poetry.group.migration.dependencies]
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
